+++ lib/cetustek.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require_relative "cetustek/version"
+require_relative "cetustek/errors"
 require_relative "cetustek/configuration"
 require_relative "cetustek/models/allowance_data"
 require_relative "cetustek/create_invoice"
@@ -12,8 +13,6 @@ require_relative "cetustek/cancel_allowance"
 require_relative "cetustek/phone_barcode"
 
 module Cetustek
-  class Error < StandardError; end
-
   class << self
     def configure
       yield(config)